refac: move UI for point clouds to separate class (#1813)
